+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+## [Unreleased]
+### Added
+- @jeremywrowe - Add 3D Secure 2 complete transaction
 
 ## [2.0.23] - 2019-08-16
 ### Added